>> merge commit 2020.0724 version 1.0.0.20191020.beta

+++ src/main/java/misaka/net/MaxMindIpGeolocationProvider.java
@@ -8,6 +8,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.math.BigDecimal;
 import java.net.InetAddress;
 
 /**
@@ -57,8 +58,8 @@ public class MaxMindIpGeolocationProvider implements IpGeolocationProvider {
 //            ipGeolocation.setDistrictCode();
 //            ipGeolocation.setStreet();
 //            ipGeolocation.setStreetCode();
-            ipGeolocation.setLatitude(location.getLatitude());
-            ipGeolocation.setLongitude(location.getLongitude());
+            ipGeolocation.setLatitude(BigDecimal.valueOf(location.getLatitude()));
+            ipGeolocation.setLongitude(BigDecimal.valueOf(location.getLongitude()));
 //            ipGeolocation.setElevation();
 //            ipGeolocation.set;
             return ipGeolocation;
